Split main.py into main.py + basics.py + api.py + web.py; introduced Flask Blueprints
