+++ static/fullscreen.less
@@ -86,7 +86,15 @@
       width: 504px;
       height: 276px;
    }
+
+   div.try-again {
+      width: auto;
+   }
    
+   div.get-help {
+      width: auto;
+   }
+
    div.board-help {
       display: block;
       z-index: 2;
