+++ templates/admin.html
@@ -3,7 +3,7 @@
 
 {% block styles %}
 
-<link href="{{ url_for('static', filename='jquery-ui.css') }}" rel="stylesheet" type="text/css">
+<link href="{{ url_for('web.static', filename='jquery-ui.css') }}" rel="stylesheet" type="text/css">
 
 <style>
