+++ templates/back.html
@@ -1,6 +1,6 @@
 
 <div class="logo-back">
- 	<a class="backlink" href="{{ url_for('main') }}">
+ 	<a class="backlink" href="{{ url_for('web.main') }}">
       	<div class="circle"><span class="glyphicon glyphicon-home" title="Aftur í aðalskjá"></span></div>
  	</a>
 </div>
