+++ templates/container-explo.html
@@ -5,5 +5,5 @@
 <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.11.2/jquery.min.js"></script>
 <script src="https://ajax.googleapis.com/ajax/libs/jqueryui/1.11.2/jquery-ui.min.js"></script>
 <!-- Touch support for jQuery -->
-<script src="{{ url_for('static', filename='jquery.ui.touch-punch.2019.js') }}"></script>
+<script src="{{ url_for('web.static', filename='jquery.ui.touch-punch.2019.js') }}"></script>
 {% endblock %}
