+++ templates/login-error.html
@@ -7,7 +7,7 @@
 <meta property="og:type" content="website" />
 <meta property="og:site_name" content="Netskrafl" />
 <meta property="og:description" content="Spilaðu skrafl á netinu" />
-<meta property="og:image" content="{{ url_for('static', filename='NetskraflFacebookImageSq.jpg', _external=True) | safe }}" />
+<meta property="og:image" content="{{ url_for('web.static', filename='NetskraflFacebookImageSq.jpg', _external=True) | safe }}" />
 <meta property="og:image:type" content="image/jpeg" />
 <meta property="og:image:width" content="600" />
 <meta property="og:image:height" content="600" />
@@ -60,12 +60,12 @@ function getCookie(name) {
 function initMain() {
    $("div.try-again").click(
       function() {
-         window.location.assign("{{ url_for('login') }}");
+         window.location.assign("{{ url_for('web.login') }}");
       }
    );
    $("div.get-help").click(
       function() {
-         window.location.assign("{{ url_for('faq') }}#faq-31");
+         window.location.assign("{{ url_for('web.faq') }}#faq-31");
       }
    );
 }
