+++ templates/logo.html
@@ -1,6 +1,6 @@
 
 <div class="logo">
- 	<a class="nodecorate" href="{{ url_for('main') }}">
-    	<img src="{{ url_for('static', filename='Netskrafl.png') }}" width=46 height=400 alt="Netskrafl">
+ 	<a class="nodecorate" href="{{ url_for('web.main') }}">
+    	<img src="{{ url_for('web.static', filename='Netskrafl.png') }}" width=46 height=400 alt="Netskrafl">
  	</a>
 </div>
