+++ templates/promo-krafla.html
@@ -1,5 +1,5 @@
 
-<img src="{{ url_for('static', filename='promo-krafla.png')}}" height="553" width="575">
+<img src="{{ url_for('web.static', filename='promo-krafla.png')}}" height="553" width="575">
 
 <div class="btn-krafla-yes" title="Loka"
    onmouseover="buttonOver(this)" onmouseout="buttonOut(this)"
