+++ templates/skilmalar.html
@@ -2,7 +2,7 @@
 {% extends "container.html" %}
 
 {% block styles %}
-<link href="{{ url_for('static', filename='jquery-ui.css') }}" rel="stylesheet" type="text/css">
+<link href="{{ url_for('web.static', filename='jquery-ui.css') }}" rel="stylesheet" type="text/css">
 {% endblock %}
 
 {% block logo %}
@@ -43,7 +43,7 @@
 
 <!-- Display user identifier in upper right corner -->
 
-<div class="userid" onclick="location.href='{{ url_for('userprefs') }}'" title="Upplýsingar um leikmann">
+<div class="userid" onclick="location.href='{{ url_for('web.userprefs') }}'" title="Upplýsingar um leikmann">
 <span class="glyphicon glyphicon-address-book"></span>&nbsp;{{ user.nickname() }}
 </div>
