+++ templates/userprefs.html
@@ -2,7 +2,7 @@
 {% extends "container.html" %}
 
 {% block styles %}
-<link href="{{ url_for('static', filename='jquery-ui.css') }}" rel="stylesheet" type="text/css">
+<link href="{{ url_for('web.static', filename='jquery-ui.css') }}" rel="stylesheet" type="text/css">
 {% endblock %}
 
 {% block scripts %}
@@ -46,7 +46,7 @@ function submitForm() {
 
    <div class="loginhdr"><span class='glyphicon glyphicon-address-book'></span> Upplýsingar um leikmann</div>
    <div>
-   <form name="frm1" id="frm1" action="{{ url_for('userprefs', from=from_url) | safe }}" method="post">
+   <form name="frm1" id="frm1" action="{{ url_for('web.userprefs', from=from_url) | safe }}" method="post">
 
       <div class="dialog-spacer"><span class="caption">Einkenni:</span>
          <input type="text" name="nickname" id="nickname" tabindex="1"
@@ -172,7 +172,7 @@ function submitForm() {
 {% if from_url %}
       onclick="location.href='{{ from_url }}'"
 {% else %}
-      onclick="location.href='{{ url_for('main') | safe }}'"
+      onclick="location.href='{{ url_for('web.main') | safe }}'"
 {% endif %}
       onmouseover="buttonOver(this)"
       onmouseout="buttonOut(this)"><span class="glyphicon glyphicon-remove"></span></div>
@@ -186,13 +186,13 @@ function submitForm() {
 {% if uf.friend %}
    <!-- Friend badge -->
    <div class="modal-close" id="user-unfriend" title="Hætta sem vinur" tabindex="11"
-      onclick="location.href='{{ url_for('friend', action=2) }}'"
+      onclick="location.href='{{ url_for('web.friend', action=2) }}'"
       onmouseover="buttonOver(this)"
       onmouseout="buttonOut(this)"><span class="glyphicon glyphicon-coffee-cup"></span>&nbsp;&nbsp;Þú ert vinur Netskrafls!</div>
 {% else %}
    <!-- Become friend button -->
    <div class="modal-close" id="user-friend" title="Gerast vinur" tabindex="11"
-      onclick="location.href='{{ url_for('friend', action=1) }}'"
+      onclick="location.href='{{ url_for('web.friend', action=1) }}'"
       onmouseover="buttonOver(this)"
       onmouseout="buttonOut(this)"><span class="glyphicon glyphicon-coffee-cup"></span>&nbsp;&nbsp;Gerast vinur Netskrafls</div>
 {% endif %}
@@ -206,11 +206,11 @@ function submitForm() {
 
 <!-- Audio resources -->
 <audio id="your-turn" preload="auto">
-   <source src="{{ url_for('static', filename='your-turn.mp3') }}" type="audio/mpeg">
+   <source src="{{ url_for('web.static', filename='your-turn.mp3') }}" type="audio/mpeg">
 </audio>
 
 <audio id="you-win" preload="auto">
-   <source src="{{ url_for('static', filename='you-win.mp3') }}" type="audio/mpeg">
+   <source src="{{ url_for('web.static', filename='you-win.mp3') }}" type="audio/mpeg">
 </audio>
 
 <script>
